requirements.txt
Flask
Flask-Core
flask-marshmallow
Flask-HTTPAuth
Flask-Dance
marshmallow-sqlalchemy
psycopg2-binary
flask_sqlalchemy
requests
requests-xml
ariadne
